chore(routes): Rename routes slots in preparation for nested dynamic route
