+++ app/dashboard/@sheet/(.)records/[recordId]/page.tsx
@@ -1,5 +1,5 @@
 import { redirect } from "next/navigation";
-import { RecordDetail } from "@/app/dashboard/records/[id]/record";
+import { RecordDetail } from "@/app/dashboard/records/[recordId]/record";
 import { auth } from "@clerk/nextjs/server";
 import { RouterSheet } from "@/components/router-sheet";
 import db from "@/db";
@@ -8,13 +8,13 @@ import { and, eq } from "drizzle-orm";
 import { notFound } from "next/navigation";
 import { Metadata } from "next";
 
-export async function generateMetadata({ params }: { params: Promise<{ id: string }> }): Promise<Metadata> {
+export async function generateMetadata({ params }: { params: Promise<{ recordId: string }> }): Promise<Metadata> {
   const { orgId } = await auth();
   if (!orgId) {
     redirect("/");
   }
 
-  const id = (await params).id;
+  const id = (await params).recordId;
 
   const record = await db.query.records.findFirst({
     where: and(eq(schema.records.clerkOrganizationId, orgId), eq(schema.records.id, id)),
@@ -29,12 +29,12 @@ export async function generateMetadata({ params }: { params: Promise<{ id: strin
   };
 }
 
-export default async function Page({ params }: { params: Promise<{ id: string }> }) {
+export default async function Page({ params }: { params: Promise<{ recordId: string }> }) {
   const { orgId } = await auth();
   if (!orgId) {
     redirect("/");
   }
-  const id = (await params).id;
+  const id = (await params).recordId;
   return (
     <RouterSheet title="Record">
       <RecordDetail clerkOrganizationId={orgId} id={id} />
