+++ app/dashboard/@sheet/(.)users/[userId]/page.tsx
@@ -1,5 +1,5 @@
 import { auth } from "@clerk/nextjs/server";
-import { UserDetail } from "@/app/dashboard/users/[id]/record-user";
+import { UserDetail } from "@/app/dashboard/users/[userId]/record-user";
 import { notFound, redirect } from "next/navigation";
 import { RouterSheet } from "@/components/router-sheet";
 import db from "@/db";
@@ -8,13 +8,13 @@ import * as schema from "@/db/schema";
 import { and, eq } from "drizzle-orm";
 import { Metadata } from "next";
 
-export async function generateMetadata({ params }: { params: Promise<{ id: string }> }): Promise<Metadata> {
+export async function generateMetadata({ params }: { params: Promise<{ userId: string }> }): Promise<Metadata> {
   const { orgId } = await auth();
   if (!orgId) {
     redirect("/");
   }
 
-  const id = (await params).id;
+  const id = (await params).userId;
 
   const user = await db.query.users.findFirst({
     where: and(eq(schema.users.clerkOrganizationId, orgId), eq(schema.users.id, id)),
@@ -29,12 +29,12 @@ export async function generateMetadata({ params }: { params: Promise<{ id: strin
   };
 }
 
-export default async function Page({ params }: { params: Promise<{ id: string }> }) {
+export default async function Page({ params }: { params: Promise<{ userId: string }> }) {
   const { orgId } = await auth();
   if (!orgId) {
     redirect("/");
   }
-  const id = (await params).id;
+  const id = (await params).userId;
   return (
     <RouterSheet title="User">
       <UserDetail clerkOrganizationId={orgId} id={id} />
