+++ app/dashboard/records/[recordId]/record-images.tsx
