+++ app/dashboard/records/[recordId]/record.tsx
